+++ R/RPPA.R
@@ -54,6 +54,8 @@
 #' - 21Q3: no change, no further releases are scheduled at this time.
 #' 
 #' - 21Q4: no change, no further releases are scheduled at this time.
+#' 
+#' - 22Q1: no change, no further releases are scheduled at this time.
 #'
 #' @docType data
 #'